added cf-turnstile to pages and enhanced features
