+++ components/Shared/CloudflareTurnstile.tsx
@@ -29,6 +29,10 @@ const CloudFlareTurnStileComp = (
 		);
 	}
 
+	if (status === "success") {
+		(ref as NonNullable<ValidRefType<typeof ref>>).current?.remove();
+	}
+
 	return (
 		<Turnstile
 			ref={ref}
