+++ pages/forgot-password/index.tsx
@@ -1,6 +1,8 @@
+import CloudflareTurnstile from "@/components/Shared/CloudflareTurnstile";
 import { Button, Input, useToast } from "@/components/ui";
 import { AuthLayout } from "@/layouts";
 import { type ForgotPasswordType, callApi, zodValidator } from "@/lib";
+import { useCloudflareTurnstile } from "@/lib/hooks";
 import { zodResolver } from "@hookform/resolvers/zod";
 import Link from "next/link";
 import { useRouter } from "next/router";
@@ -9,6 +11,8 @@ import { useForm } from "react-hook-form";
 const ForgotPasswordPage = () => {
 	const router = useRouter();
 	const { toast } = useToast();
+	const { cfTurnStile, checkBotStatus, handleBotStatus } =
+		useCloudflareTurnstile();
 
 	const {
 		handleSubmit,
@@ -53,7 +57,9 @@ const ForgotPasswordPage = () => {
 			<form
 				onSubmit={(event) => {
 					event.preventDefault();
-					void handleSubmit(onSubmit)(event);
+					const response = checkBotStatus();
+
+					response && void handleSubmit(onSubmit)(event);
 				}}
 				className="flex flex-col gap-5"
 			>
@@ -88,7 +94,10 @@ const ForgotPasswordPage = () => {
 						)}
 					</div>
 				</div>
-
+				<CloudflareTurnstile
+					ref={cfTurnStile}
+					onStatusChange={handleBotStatus}
+				/>
 				<div className="flex flex-col gap-4 text-center">
 					<Button
 						disabled={isSubmitting}
